Support JB customer pricing module, drop support for Mageplaza better tier pricing
