+++ src/Actions/UpdateCustomerPrices.php
@@ -7,7 +7,7 @@ use JustBetter\MagentoCustomerPrices\Contracts\UpdatesMagentoCustomerPrices;
 use JustBetter\MagentoCustomerPrices\Data\CustomerPriceData;
 use JustBetter\MagentoCustomerPrices\Models\MagentoCustomerPrice;
 
-class UpdateMageplazaCustomerPrices implements UpdatesMagentoCustomerPrices
+class UpdateCustomerPrices implements UpdatesMagentoCustomerPrices
 {
     public function __construct(protected Magento $magento)
     {
@@ -15,30 +15,16 @@ class UpdateMageplazaCustomerPrices implements UpdatesMagentoCustomerPrices
 
     public function update(MagentoCustomerPrice $model): void
     {
-        $recordId = 0;
-        $mageplazaData = $model->getDataCollection()->map(function (CustomerPriceData $price) use (&$recordId) {
-            $data = $price->toMageplazaData();
-
-            // Retrieve the customer name from Magento
-            $customer = $this->magento->get('customers/'.$price->getCustomerId())->json();
-
-            $data['customer'] = implode(' ', [$customer['firstname'], $customer['lastname']]);
-            $data['record_id'] = (string) $recordId;
-
-            $recordId++;
-
-            return $data;
+        $prices = $model->getDataCollection()->map(function (CustomerPriceData $price) {
+            return [
+                'customer_id' => $price->customerId,
+                'quantity' => $price->quantity,
+                'price' => $price->price->getAmount()->toFloat(),
+            ];
         })->toArray();
 
-        $response = $this->magento->put("products/$model->sku", [
-            'product' => [
-                'custom_attributes' => [
-                    [
-                        'attribute_code' => 'mp_specific_customer',
-                        'value' => json_encode($mageplazaData),
-                    ],
-                ],
-            ],
+        $response = $this->magento->post('customer-pricing/'.urlencode($model->sku), [
+            'customerPrices' => $prices,
         ]);
 
         if (! $response->ok()) {
